Start adding test for symmetric tensor contraction
